Move demo scenes to common directory
+ set module resolution to 'node'
+ merge lineDemo and ellipseDemo to staticDemo
