+++ demo/src/scene/index.ts
@@ -1,0 +1,3 @@
+export * from './animationDemo'
+export * from './staticDemo'
+export * from './scalingDemo'
