+++ demo/src/scene/scalingDemo.ts
