+++ demo/src/scene/staticDemo.ts
@@ -1,9 +1,16 @@
+import { ellipse } from 'Canvas/index'
 import { line } from 'Canvas/line'
 import { CanvasInfo } from 'Canvas/types'
 
-import boundaryBox from './boundaryBox'
+import boundaryBox from '../boundaryBox'
 
 const boundary = 10
+
+function staticDemo(ci: CanvasInfo) {
+  boundaryBox(ci, boundary)
+  lineDemo(ci)
+  ellipseDemo(ci)
+}
 
 function lineDemo(ci: CanvasInfo) {
   boundaryBox(ci, boundary)
@@ -27,4 +34,15 @@ function lineDemo(ci: CanvasInfo) {
   })
 }
 
-export { lineDemo }
+function ellipseDemo(ci: CanvasInfo) {
+  for (let r = 1; r < 5; r++) {
+    ellipse(ci.ctx, {
+      x: ci.width / 2,
+      y: ci.height / 2,
+      rx: (ci.width / 2 / 5) * r,
+      ry: (ci.height / 2 / 7) * r,
+      color: 'orange',
+    })
+  }
+}
+export { staticDemo }
